Fixed Issue 327 - Maximum value in Monitor charts is Integer.MAX_VALUE  - replaced incorrect casting to int by cast to long
git-svn-id: https://svn.java.net/svn/visualvm~svn/trunk@2283 ba0df5e5-9053-6d9f-941f-f112716aa91a

+++ visualvm/charts/src/com/sun/tools/visualvm/charts/xy/SimpleXYChartUtils.java
@@ -413,7 +413,7 @@ public class SimpleXYChartUtils {
             public String getRowValue(int index, long itemValue) {
                 double value = itemValue * chartFactor;
                 switch (chartType) {
-                    case TYPE_BYTES  : return formatBytes((int)value);
+                    case TYPE_BYTES  : return formatBytes((long)value);
                     case TYPE_PERCENT: return formatPercent(value);
                     default:           return formatDecimal(value);
                 }
